requirements.txt
aiohttp==3.8.5
asyncio==3.4.3
async-timeout==4.0.2
attrs==23.1.0
chardet==5.2.0
discord.py==2.3.1
idna==3.4
multidict==6.0.4
typing-extensions==4.7.1
yarl==1.9.2
requests==2.31.0
openai==0.27.8
python-dotenv==1.0.0
pytest==7.4.0
Pillow==10.0.0
replicate==0.11.0